+++ src/easyscience/Fitting/__init__.py
@@ -1,4 +1,3 @@
 from .fitter import Fitter  # noqa: F401, E402
-from .minimizers.fitting_base import FitResults  # noqa: F401, E402
-# Causes a circular import
-# from .multi_fitter import MultiFitter  # noqa: F401, E402
+from .minimizers.minimizer_base import FitResults  # noqa: F401, E402
+from .multi_fitter import MultiFitter  # noqa: F401, E402
